There is still the name from the original licence file in here
The place I've copied it from can be found here:
https://github.com/jokesterfr/node-pcsc/tree/master/src

Though I've took a similar project structure and
copied over the license and it was a good project
to learn nodejs C++ Addon API, I will replace the
name by mine for this project due to the fact that
the code is 100% new. If anyone object please tell.